First real commit.  Builds clean
This is a first-cut of version 2.2.0, the first
set of work this has seen in quite a while.

- Moved things around a bit to make it more
maintainable.
- Made the build system work with CMake to
make everyone's life a LOT easier.
- Brought in the runtime library piece for
runit and renamed it in the build for now
as this is a dependency OTHERWISE and we're
trying to limit dependencies for now.

This will need some smoke test before it's
tagged as a 2.2.0 release, but it's looking
like maybe we've got this.